Added web service

+++ Sharpen/Extensions.cs
@@ -608,7 +608,7 @@ namespace Sharpen
 			return val;
 		}
 
-		public static int GetTotalInFixed (this Inflater inf)
+		public static long GetTotalInFixed (this Inflater inf)
 		{
 			if (inf.TotalIn > 0)
 				return inf.TotalIn + 4;
